+++ include/H2DE/H2DE_window.h
@@ -15,6 +15,7 @@ private:
     void create();
     void initSettings() const;
     void quitSDL() const;
+    SDL_WindowFlags getFlags(bool fullscreen, bool resizable) const;
 
 public:
     H2DE_Window(H2DE_Engine* engine, H2DE_WindowData data);
@@ -22,8 +23,20 @@ public:
 
     void saveState() const;
 
-    SDL_Window* getWindow() const;
-    SDL_Renderer* getRenderer() const;
+    friend H2DE_AbsPos H2DE_GetWindowPos(const H2DE_Engine* engine);
+    friend H2DE_AbsSize H2DE_GetWindowSize(const H2DE_Engine* engine);
+
+    friend void H2DE_SetWindowPos(const H2DE_Engine* engine, const H2DE_AbsPos& pos);
+    friend void H2DE_SetWindowSize(const H2DE_Engine* engine, const H2DE_AbsSize& size);
+    friend void H2DE_SetWindowMinimumSize(const H2DE_Engine* engine, const H2DE_AbsSize& minimumSize);
+    friend void H2DE_SetWindowMaximumSize(const H2DE_Engine* engine, const H2DE_AbsSize& maximumSize);
+    friend void H2DE_SetWindowTitle(const H2DE_Engine* engine, const std::string& title);
+    friend void H2DE_SetWindowIcon(const H2DE_Engine* engine, const std::string& textureName);
+    friend void H2DE_SetWindowFullscreen(const H2DE_Engine* engine, bool fullscreen);
+    friend void H2DE_SetWindowResizable(const H2DE_Engine* engine, bool resizable);
+    friend void H2DE_SetWindowGrab(const H2DE_Engine* engine, bool grab);
+
+    friend class H2DE_Engine;
 };
 
 #endif